requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
tqdm>=4.65.0
tensorflow-datasets>=4.9.0
tensorflow>=2.12.0
